refactor: reorganize Compose components into dedicated directory
- Create /src/components/compose-components/ for all Compose-related components
- Rename components for clarity:
  - ComposeComponentForm → ComponentCreationForm
  - ComposePreview → ComponentDetailView
  - Keep ComposeComponentGrid name
- Move kotlin-playground components into compose-components/
- Remove redundant ComponentDetailClient wrapper
- Add comprehensive documentation and README files
- Update all imports across the codebase
- Add barrel exports for convenient imports

This improves code organization, makes components easier to discover,
and provides clearer naming that reflects each component's purpose.
