+++ docs/compose-components-reorganization-summary.md
@@ -1,0 +1,91 @@
+# Compose Components Reorganization Summary
+
+## What Was Changed
+
+### Before
+Components were scattered in the root of `/src/components/`:
+```
+src/components/
+├── compose-component-form.tsx
+├── compose-component-grid.tsx
+├── compose-preview.tsx
+├── kotlin-playground.tsx
+└── kotlin-playground/
+    ├── kotlin-compose-editor.tsx
+    └── kotlin-compose-preview.tsx
+```
+
+### After
+All Compose-related components are now organized under `/src/components/compose-components/`:
+```
+src/components/
+└── compose-components/
+    ├── component-creation-form.tsx    # Was: compose-component-form.tsx
+    ├── component-detail-view.tsx      # Was: compose-preview.tsx
+    ├── component-grid.tsx             # Was: compose-component-grid.tsx
+    ├── kotlin-playground.tsx          # Base playground component
+    ├── kotlin-playground/             # Specialized playground components
+    │   ├── kotlin-compose-editor.tsx
+    │   ├── kotlin-compose-preview.tsx
+    │   └── README.md
+    ├── index.ts                       # Barrel exports
+    └── README.md                      # Component documentation
+```
+
+## Component Renames
+
+| Old Name | New Name | Purpose |
+|----------|----------|---------|
+| `ComposeComponentForm` | `ComponentCreationForm` | Form for creating new components |
+| `ComposePreview` | `ComponentDetailView` | Full view of component with preview and code |
+| `ComposeComponentGrid` | `ComposeComponentGrid` | Grid layout (kept same name) |
+
+## Import Updates
+
+### Old Imports
+```typescript
+import { ComposeComponentForm } from '@/components/compose-component-form'
+import { ComposePreview } from '@/components/compose-preview'
+import { ComposeComponentGrid } from '@/components/compose-component-grid'
+import { KotlinComposeEditor } from '@/components/kotlin-playground/kotlin-compose-editor'
+```
+
+### New Imports
+```typescript
+import { ComponentCreationForm } from '@/components/compose-components/component-creation-form'
+import { ComponentDetailView } from '@/components/compose-components/component-detail-view'
+import { ComposeComponentGrid } from '@/components/compose-components/component-grid'
+import { KotlinComposeEditor } from '@/components/compose-components/kotlin-playground/kotlin-compose-editor'
+```
+
+### Using Barrel Imports
+```typescript
+import { 
+  ComponentCreationForm,
+  ComponentDetailView,
+  ComposeComponentGrid,
+  KotlinComposeEditor,
+  KotlinComposePreview
+} from '@/components/compose-components'
+```
+
+## Benefits
+
+1. **Better Organization**: All Compose-related components are in one place
+2. **Clearer Naming**: Component names better reflect their purpose
+3. **Easier Discovery**: Developers can find all related components in one directory
+4. **Cleaner Root**: The `/src/components/` directory is less cluttered
+5. **Consistent Structure**: Similar components are grouped together
+
+## Updated Files
+
+- All page files that imported these components
+- Component documentation headers
+- README files
+- Component guides
+
+## No Breaking Changes
+
+- All functionality remains the same
+- Only file locations and component names changed
+- All imports have been updated automatically
