+++ docs/kotlin-playground-review-summary.md
@@ -1,0 +1,88 @@
+# Kotlin Playground Integration - Code Review Summary
+
+## Current State
+
+### Components Structure
+```
+src/
+├── components/
+│   ├── kotlin-playground/              # Kotlin-specific components
+│   │   ├── index.ts                    # Barrel export
+│   │   ├── kotlin-compose-editor.tsx   # Editor wrapper for Compose
+│   │   └── kotlin-compose-preview.tsx  # Preview-only component
+│   ├── compose-preview.tsx             # High-level preview component
+│   ├── compose-component-form.tsx      # Form using editor + preview
+│   └── kotlin-playground.tsx           # Base playground component
+├── hooks/
+│   └── useKotlinPlayground.ts          # Hook for loading playground
+└── types/
+    └── kotlin-playground.d.ts          # TypeScript definitions
+```
+
+### Component Usage
+
+1. **KotlinComposeEditor** - Used for editable code views
+   - Used in: `compose-component-form.tsx`, `compose-preview.tsx`
+   - Features: Optionally hide run button, Compose-specific defaults
+
+2. **KotlinComposePreview** - Used for preview-only views
+   - Used in: `compose-component-form.tsx`, `compose-preview.tsx`
+   - Features: Auto-compile, hidden editor, compilation callbacks
+
+3. **ComposePreview** - High-level component for component detail pages
+   - Used in: `component-detail-client.tsx`
+   - Features: Preview + read-only code, copy/download buttons
+
+4. **KotlinPlayground** - Base component (dependency for specialized components)
+   - Not used directly, but required by editor/preview components
+   - Features: Full playground functionality, code change detection
+
+### Key Features Implemented
+
+1. **Tight Preview/UI State Coupling**
+   - Create button only enabled when preview is fully loaded
+   - Screenshot capture capability for component cover photos
+   - Proper compilation state management
+
+2. **Specialized Components**
+   - Separate editor and preview components for different use cases
+   - Compose-wasm platform hardcoded for consistency
+   - Flexible configuration options
+
+3. **Code Organization**
+   - Clean separation of concerns
+   - Proper TypeScript types
+   - Reusable hooks
+
+### Recent Fixes
+
+1. Fixed infinite compilation loop by:
+   - Using refs to track compilation state
+   - Preventing component remounting with CSS display toggle
+   - Proper dependency management in useEffect
+
+2. Fixed TypeScript compilation errors:
+   - Resolved circular reference in setInterval
+   - Fixed missing KotlinPlayground import
+   - Proper type annotations
+
+3. Fixed code change detection:
+   - Polling approach for CodeMirror changes
+   - Multiple fallback methods
+   - Proper cleanup on unmount
+
+### Documentation Files
+
+- `kotlin-playground-integration-plan.md` - Original implementation plan
+- `kotlin-playground-limitations.md` - Known limitations and workarounds
+- Example files for testing components
+
+## Build Status
+
+✅ Build successful
+✅ No TypeScript errors
+⚠️  Minor ESLint warning about ref in cleanup (acceptable)
+
+## Next Steps
+
+The Kotlin Playground integration is complete and ready for use. All components are properly organized, typed, and tested.
