+++ src/components/compose-components/index.ts
@@ -1,0 +1,10 @@
+// Component exports for Compose-related components
+
+// High-level components
+export { ComponentCreationForm } from './component-creation-form'
+export { ComponentDetailView } from './component-detail-view'
+export { ComposeComponentGrid } from './component-grid'
+
+// Kotlin Playground components
+export { KotlinComposeEditor } from './kotlin-playground/kotlin-compose-editor'
+export { KotlinComposePreview } from './kotlin-playground/kotlin-compose-preview'
